# backend/OpenAPI_docs.yaml
openapi: 3.0.0
info:
  title: Kanban Board API
  version: 1.0.0
  description: API для управления канбан-доской (аналог Trello)
servers:
  - url: http://localhost:3001
    description: Локальный сервер
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Произошла ошибка"
    ValidationErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Поле title обязательно"
...
    SUserRegister:
      type: object
      required: [username, email, password, password_confirm]
      properties:
        username: { type: string }
        email: { type: string, format: email }
        password: { type: string }
        password_confirm: { type: string }
      example:
        username: "john_doe"
        email: "john@example.com"
        password: "securepass"
        password_confirm: "securepass"
    SUserLogin:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
      example:
        email: "john@example.com"
        password: "securepass"
    SUserUpdate:
      type: object
      properties:
        username: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        avatar_id: { type: integer, nullable: true }
        bio: { type: string, nullable: true, maxLength: 500 }
        birth_date: { type: string, format: date, nullable: true }
        gender: { type: string, enum: [male, female], nullable: true }
      example:
        username: "new_name"
        bio: "Привет, я новый пользователь"
    SUser:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string, format: email }
        avatar_id: { type: integer, nullable: true }
        bio: { type: string, nullable: true }
        birth_date: { type: string, format: date, nullable: true }
        gender: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
      example:
        id: 1
        username: "john_doe"
        email: "john@example.com"
        avatar_id: null
        bio: null
        birth_date: null
        gender: null
        created_at: "2024-01-01T12:00:00Z"
    SUserListResponse:
      type: object
      properties:
        users: { type: array, items: { $ref: "#/components/schemas/SUser" } }
        next_cursor: { type: string, nullable: true }
        previous_cursor: { type: string, nullable: true }
    RegisterResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        user_id: { type: integer, example: 1 }
        message: { type: string, example: "Регистрация прошла успешно" }
    LoginResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string, example: "Вы вошли в аккаунт" }
        access_token: { type: string, example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." }
        refresh_token: { type: string, example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." }
        token_type: { type: string, example: "bearer" }
    RefreshResponse:
      type: object
      properties:
        access_token: { type: string, example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." }
        token_type: { type: string, example: "bearer" }
    LogoutResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
    # Boards
...
security:
  - Bearer: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SUserRegister" }
      responses:
        "201":
          description: Успешная регистрация
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegisterResponse" }
        "400":
          description: Ошибка валидации
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationErrorResponse" }
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /auth/login:
    post:
      tags: [Auth]
      summary: Вход в систему
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SUserLogin" }
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "400":
          description: Неверный email или пароль
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Обновление access токена
      parameters:
        - name: refresh_token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Новый access токен
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RefreshResponse" }
        "400":
          description: Неверный refresh токен
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Выход из системы
      security:
        - Bearer: []
      responses:
        "200":
          description: Успешный выход
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LogoutResponse" }
        "401":
          description: Не авторизован
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /auth/me:
    get:
      tags: [Auth]
      summary: Получить информацию о текущем пользователе
      security:
        - Bearer: []
      responses:
        "200":
          description: Данные пользователя
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SUser" }
        "401":
          description: Не авторизован
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    patch:
      tags: [Auth]
      summary: Обновить профиль текущего пользователя
      security:
        - Bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SUserUpdate" }
      responses:
        "200":
          description: Обновлённые данные пользователя
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SUser" }
        "400":
          description: Ошибка валидации
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationErrorResponse" }
        "401":
          description: Не авторизован
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Файл не найден (при обновлении аватарки)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /auth/users:
    get:
      tags: [Auth]
      summary: Список пользователей с пагинацией
      security:
        - Bearer: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: direction
          in: query
          schema:
            type: string
            enum: [after, before]
            default: after
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        "200":
          description: Список пользователей
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SUserListResponse" }
        "401":
          description: Не авторизован
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /auth/users/by-email:
    get:
      tags: [Auth]
      summary: Найти пользователя по email
      security:
        - Bearer: []
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
          example: "john@example.com"
      responses:
        "200":
          description: Найденный пользователь
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SUser"
        "400":
          description: Параметр email обязателен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/users/{user_id}:
    get:
      tags: [Auth]
      summary: Получить пользователя по ID
      security:
        - Bearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Найденный пользователь
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SUser"
        "401":
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...